backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-security-too==5.6.0
flask-cors==4.0.0
flask-migrate==4.0.7
argon2-cffi==23.1.0
email-validator==2.1.0
